rename dir. update selected fingering button styling.
